configure with ts
