+++ src/components/App.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import MovieCard from "./MovieCard";
 
 function App() {
