+++ src/components/MovieCard.tsx
@@ -1,0 +1,21 @@
+interface Props {
+  title: string;
+  posterUrl: string;
+  genres: string[];
+}
+
+function MovieCard({
+  title,
+  posterUrl = "https://m.media-amazon.com/images/M/MV5BOTJjNzczMTUtNzc5MC00ODk0LWEwYjgtNzdiOTEyZmQxNzhmXkEyXkFqcGdeQXVyNzMzMjU5NDY@._V1_UY268_CR1,0,182,268_AL_.jpg",
+  genres,
+}: Props) {
+  return (
+    <div className="movie-card">
+      <img src={posterSrc} alt={title} />
+      <h2>{title}</h2>
+      <small>{genres.join(", ")}</small>
+    </div>
+  );
+}
+
+export default MovieCard;
